:green_heart: :recycle: Extract class_proc_test.go
